Refactor tool_shed.util.container_util into galaxy.
